pyproject.toml
[project]
name = "sdev265-group4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[tool.ruff]
fix = true
line-length = 120
#"full" | "concise" | "grouped" | "json" | "junit" | "github" | "gitlab" | "pylint" | "azure"
output-format = "full"
show-fixes = true

# [tool.ruff.per-file-target-version]
# Override the project-wide Python version for a developer scripts directory:
# "scripts/*.py" = "py312"

[tool.ruff.format]
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true
